coins and courses implemented

+++ src/Components/Courses.js
@@ -8,26 +8,91 @@ import { useNavigate } from 'react-router-dom';
 import Modal from './Modal';
 import { updateUser } from '../Redux/Reducers/userReducer';
 
-import { setCourse, setEdit, setLogout, setMyCourses, setUploadPix } from '../Redux/Reducers/generalReducer';
+import { setCourse, setEdit, setLogout, setMyCourses, setaddCourse } from '../Redux/Reducers/generalReducer';
 import { funSeque } from 'flame-tools';
 import UploadPix from './UploadPix';
 import axios from 'axios';
-import { setload } from '../Redux/Reducers/displayReducer';
+import { setalert, setload } from '../Redux/Reducers/displayReducer';
+import { ProcessManager } from '../Process';
 
 
 
 function Courses() {
   const navFall=useSelector((state)=>state.displayReducer.display.navigationFall);
   const user=useSelector((state)=>state.userReducer.user);
-  const {edit, logout, uploadPix, myCourses}=useSelector((state)=>state.generalReducer.general);
+  const { addCourse, uploadPix, myCourses, network}=useSelector((state)=>state.generalReducer.general);
+ 
   const courses=useSelector((state)=>state.generalReducer.general.courses);
   const [pereson, setPerson]=useState({...user});
   const [campuses, setcampuses] = useState([]);
+  const {alert, locked}=useSelector((state)=>state.displayReducer.display)
+  useEffect(()=>{
+    localStorage.setItem('last_page', location.hash)
+    
+    }, [])
+    useEffect(()=>{
+
+      if(locked){
+       navigate('/resume')}
+     
+     }, [locked])
   const [searchTeerm, setsearchTeerm] = useState("");
   const [departments, setdepartments] = useState([]);
   let added=false;
   let dispatch=useDispatch()
   let navigate=useNavigate();
+const [level, setlevel] = useState('');
+const [department, setdepartment] = useState('');
+const [code, setcode] = useState('');
+const [title, settitle] = useState('');
+
+
+
+const createCourse=({code, title, department, level})=>{
+
+  if(network){
+    axios.post('http://192.168.43.31:5000/courses/', {code, title, campus:user.campus, department, level, user:user.email}).then((res)=>{
+      if(res.data.success){
+        dispatch(setalert({...alert, msg:res.data.message, type:'success', status:true, cap:'Success'}))
+        dispatch(setaddCourse(false))
+      }
+      else{
+        dispatch(setalert({...alert, msg:res.data.message, type:'danger', status:true, cap:'Error'}))
+      }
+    })
+  }else{
+  let feedback='';
+  feedback=ProcessManager.addProcess(
+        ()=>{
+      axios.post('http://192.168.43.31:5000/courses/', {code, title, campus:user.campus, department, level, user:user.email}).then((res)=>{
+        if(res.data.success){
+          dispatch(setalert({...alert, msg:res.data.message, type:'success', status:true, cap:'Success'}))
+        }
+        else{
+          dispatch(setalert({...alert, msg:res.data.message, type:'danger', status:true, cap:'Error'}))
+        }
+      })
+
+}
+        )
+
+        dispatch(setalert({...alert, msg:feedback, type:'bad_network', status:true, cap:'Processor'}))
+        dispatch(setaddCourse(false))
+       
+}
+}
+
+
+const postCourse=()=>{
+  axios.post('http://192.168.43.31:5000/courses/', {code, title, campus:user.campus, department, level, user:user.email}).then((response)=>{
+
+  })
+}
+useEffect(()=>{
+dispatch(setalert({...alert, msg:'Select the courses for the Semester as features/services depends on your courses', cap:'Info', status:true, type:'info'}))
+
+
+}, [])
 
   useEffect(()=>{
     axios.get('http://192.168.43.31:5000/courses/'+user.campus+'').then((response)=>{
@@ -38,6 +103,11 @@ function Courses() {
         
     }
     })
+    axios.get('http://192.168.43.31:5000/departments').then((response)=>{
+  if(response.data.success){
+    setdepartments(response.data.data)
+  }
+})
   })
 
   
@@ -55,26 +125,66 @@ function Courses() {
   </div>
 </div>
 <div className='card-body'>
+
 <div className='row'>
 <div className='col-sm-3'>
     <ul className=' list-group'>
-        <h6>Selected Semester Courses</h6>
+        <h6>Selected Semester Courses [{myCourses.length  }]</h6>
 {myCourses.map((course, key)=>{
 
         return <li key={key} className='list-group-item d-flex justify-content-between align-items-center'>
 <strong>{course.course}</strong>
 <button className='btn text-danger' onClick={()=>{
+  if(network){
     axios.delete('http://192.168.43.31:5000/mycourses/'+course.id).then((response)=>{
-        axios.get('http://192.168.43.31:5000/mycourses/'+sessionStorage.getItem('email')).then((response)=>{
+      if(response.data.success){
+        dispatch(setalert({...alert, msg:response.data.message, type:'success', status:true, cap:'Success'}))
+     
+      }else{
+        dispatch(setalert({...alert, msg:response.data.message, type:'danger', status:true, cap:'Error'}))
+     
+      } 
+     
+      axios.get('http://192.168.43.31:5000/mycourses/'+localStorage.getItem('email')).then((response)=>{
 
-        if(response.data.success){
-        
-          dispatch(setMyCourses(response.data.data))
-        
-        }
-        })
+      if(response.data.success){
+      
+        dispatch(setMyCourses(response.data.data))
+      
+      }
+      
+      })
 
-    })
+  })
+  }else{
+ dispatch(setalert({...alert, msg: ProcessManager.addProcess(()=>{
+  axios.delete('http://192.168.43.31:5000/mycourses/'+course.id).then((response)=>{
+    if(response.data.success){
+      dispatch(setalert({...alert, msg:response.data.message, type:'success', status:true, cap:'Success'}))
+   
+    }else{
+      dispatch(setalert({...alert, msg:response.data.message, type:'danger', status:true, cap:'Error'}))
+   
+    }    
+  axios.get('http://192.168.43.31:5000/mycourses/'+localStorage.getItem('email')).then((response)=>{
+
+      if(response.data.success){
+      
+        dispatch(setMyCourses(response.data.data))
+      
+      }
+   
+     
+      })
+
+  })
+
+}), type:'bad_network', status:true, cap:'Processor'}))
+}
+
+ 
+
+
 
 }}> <FontAwesomeIcon icon={faTimes}></FontAwesomeIcon> </button>
         </li>
@@ -87,11 +197,81 @@ function Courses() {
 
 
     <div className='col-sm-9'>
-<span style={{display:'flex', justifyContent:'center', alignItems:'center'}}>    <FontAwesomeIcon icon={faSearch}></FontAwesomeIcon> <input type={'search'} placeholder={' Search Courses'} className=' form-control-plaintext' onChange={(e)=>{
+    { addCourse?<Modal config={{align:'flex-end', justify:'right'}} header={'Create a New Course'} body={<>
+<div className='row'>
+
+  <div className='col-sm-6'>
+<input className='form-control' placeholder='Course Code' onChange={(e)=>{
+setcode(e.target.value)
+
+        }} />
+  </div>
+
+  <div className='col-sm-6'>
+  <select  className='form-select' onChange={(e)=>{
+setlevel(e.target.value)
+        }}>
+       <option>Select level</option>
+        <option value='100'>
+          100
+        </option>
+        <option value='200'>
+          200
+        </option>
+        <option value='300'>
+          300
+        </option>
+        <option value='400'>
+          400
+        </option>
+
+        <option value='500'>
+          500
+        </option>
+        </select>
+
+  </div>
+
+  <div className='col-sm-6'>
+<input className='form-control' placeholder='Course Title' onChange={(e)=>{
+settitle(e.target.value)
+        }} />
+  </div>
+
+  <div className='col-sm-6'>
+  <select  className='form-select' onChange={(e)=>{
+setdepartment(e.target.value)
+        }}>
+          <option>Select department</option>
+            {
+  departments.map((departmen)=>{
+return    <option value={`${departmen.name}`}>
+   {`${departmen.name}`}
+    </option>
+  })
+}
+
+        </select>
+
+     
+  </div>
+  </div>
+      </>} footer={<><button className='btn microskool-button' onClick={()=>{
+   
+      createCourse({code, title, department, level})}}>Add</button><button className='btn-close' onClick={()=>dispatch(setaddCourse(false))}></button></>} />:<></>}
+
+
+
+<span style={{display:'flex', justifyContent:'center', alignItems:'center'}}>    <FontAwesomeIcon icon={faSearch}></FontAwesomeIcon> <input type={'search'} placeholder={' Search Courses ['+courses.length+']'} className=' form-control-plaintext' onChange={(e)=>{
     setsearchTeerm(e.target.value);
     
     }}/>
-    <button className='btn btn-primary' title='New Course'> <FontAwesomeIcon icon={faAdd}></FontAwesomeIcon> </button>
+    <button className='btn microskool-button' title='New Course' onClick={()=>{
+
+     dispatch(setaddCourse(true))
+     dispatch(setalert({...alert, msg:'Ensure you do not enter misleading information into Microskool system to avoid account suspension.', cap:'Warning', status:true, type:'warning'}))
+
+    }}> <FontAwesomeIcon icon={faAdd}></FontAwesomeIcon> </button>
 </span>
 <ul className=' list-group' style={{height:'60vh',fontSize:'small', overflow:'auto'}}>
 {courses.filter((course) => { if(searchTeerm===""){
@@ -114,10 +294,18 @@ myCourses.forEach((cur)=>{
 <span>{course.title}</span>
 <span>{course.department}</span>
 {
-  added?<button className='btn text-success'><FontAwesomeIcon icon={faCheck}></FontAwesomeIcon></button>: <button className='btn text-primary' onClick={()=>{
- 
+  added?<button className='btn text-success'><FontAwesomeIcon icon={faCheck}></FontAwesomeIcon></button>: <button className='btn text-microskool' onClick={()=>{
+
+    if(network){
     axios.post('http://192.168.43.31:5000/mycourses', {user:user.email, code:course.code}).then((response)=>{
-        axios.get('http://192.168.43.31:5000/mycourses/'+sessionStorage.getItem('email')).then((response)=>{
+    if(response.data.success){
+      dispatch(setalert({...alert, msg:response.data.message, type:'success', status:true, cap:'Success'}))
+   
+    }else{
+      dispatch(setalert({...alert, msg:response.data.message, type:'danger', status:true, cap:'Error'}))
+   
+    } 
+    axios.get('http://192.168.43.31:5000/mycourses/'+localStorage.getItem('email')).then((response)=>{
 
         if(response.data.success){
         
@@ -125,12 +313,38 @@ myCourses.forEach((cur)=>{
         
         }
         })
-    })
+    })}
+    else{
+     dispatch(setalert({...alert, msg:ProcessManager.addProcess(
+      ()=>{
+      axios.post('http://192.168.43.31:5000/mycourses', {user:user.email, code:course.code}).then((response)=>{
+      if(response.data.success){
+        dispatch(setalert({...alert, msg:response.data.message, type:'success', status:true, cap:'Success'}))
+     
+      }else{
+        dispatch(setalert({...alert, msg:response.data.message, type:'danger', status:true, cap:'Error'}))
+     
+      }    
+     axios.get('http://192.168.43.31:5000/mycourses/'+localStorage.getItem('email')).then((response)=>{
+
+        if(response.data.success){
+        
+          dispatch(setMyCourses(response.data.data))
+        
+        }
+  
+       
+        })
+    })})
+        
+        , type:'bad_network', status:true, cap:'Processor'}))
+
+    }
 }} >Add</button>
 }
         </li>
-
-    })
+ 
+    }) 
 }
 </ul>
